Pachyderm 2.6.0.25 - Cluster Sources - Dom Bassett HRTF precision fix (thank you) - Improved convergence feedback - Aurralization improvements - pulse size, and pulse to sinc. - Higher Order Ambisonics - Environmental Noise Sources - improved. - removed circumsphere... not needed - Various bug fixes

+++ Pachyderm_Acoustic_Universal/Classes_HRTF_Sofa.cs
@@ -513,7 +513,7 @@ namespace Pachyderm_Acoustic
                 }
 
                 // Allocate a buffer for the data
-                double[,,] dataIR;
+                float[,,] dataIR;
 
                 try
                 {
@@ -524,7 +524,7 @@ namespace Pachyderm_Acoustic
 
                     H5S.get_simple_extent_dims(dataspaceId, dims, null);
 
-                    dataIR = new double[dims[0], dims[1], dims[2]];
+                    dataIR = new float[dims[0], dims[1], dims[2]];
                     GCHandle handle = GCHandle.Alloc(dataIR, GCHandleType.Pinned);
                     try
                     {
